Modified the "LawsMVP.sol" contract to be compatible with the construct of the "HowContract.sol", added a faked way to advance time, and performed some clean ups.
The passing of time is not simulated in the EVM, and ChatGPT advocates a
faked handlign of time. I chose to use a "one time action" indicated by the
bool variable "passTime", which is set to "true" by the constructor. Then,
the first interaction of the contract that accesses the modified will set
this to "false", such that subsequent interactions that use the modfier
will appear as if they are "after 90 days". (This is purely for the purpose
of working with the "HowContract" when using the Remix EVM.) Use of this
hack is marked with comments as "FLOW CONTROL".

The "Law" struct is now endowed with the "how" member, which is a contract
in the "HowContract.sol" (not included in this commit). Once the 90 days
"dictatorial" unilateral activity of the "owner" address has passed, any
law added/deleted/edited will need to adhere to the logic in this contract.
As far as I can tell, deployment of this contract for a given Law struct
is done at insertion of the struct in the "laws" map. The constructor needs
an argument, which is the IERC20 token that is to be used for voting logic.

At this commit, the LawsMVP.sol contract is deployable, and accepts one law
to be added; the "HowContract" functionality does not work.
IN 2023/06/03