Added a series of contracts to build the self-referential abstraction for fundamentally managing governance. The primary contract makes use of another contract that abstracts the logic of how decisions are made, including how the decision-making process itself is updated/changed. Details are desribed below, in the context of the contracts and files that are added.
Three files have been added, which represent laws with a dictatorial entity
in the decision making process, which can be changed by the dictator only.
The files are:

 "IN.sol" - fundamental laws and governanec of the entity
 "INhow_interface.sol" - an interface to a contract for decision-making
 "INhow_default.sol" - a concrete implementation of monarchy decision-making

Details on "IN.sol":

The "IN.sol" contract holds three members as its state (two members were
added for flow-control while debugging). The members are:

 (address) "founderAddr"
   This stores the address of the founder (which can be made immutable, as
   there is presently no mechanism or needs to change it). This addrtess is
   set by the constructor at instantiation.

 (unit256) "creationTimestamp"
   This stores an immutable timestamp on the blockchain when the contract
   was deployed.

 (address) "decideContractAddr"
   This is an address to a contract that is meant to be used for making
   decisions via interactions with its methods. (The idea is that this
   address can change, to point to a different contract for decisions, and
   this provides an abstraction for dynamically changing the governance
   mechanicsm).

A struct for a "law" acts as placeholder -- as per the LawsMVP.sol -- to be
changed in the future. A mapping for indices to law structs is also in place
for storing the loaws.

The constructor of this contract takes an address, which is supposed to be
the address in posession of a person, who becomes the founder, and takes
control of the entity's decision making (mechanicsm described below). The
initial "DecisionContract", whose address is to be stored in the address
member "decideContractAddr", is deployed by the constructor. This contract
(described below) is known via its concrete implementation that is known via
the file "INhow_default.sol", and it represents a dictatorial decision
making process.

The method "setDecisionContract()" is used to change the decision making
process, and therefore the governance mechanism, by changing the address
of the internal member "decideContractAddr" to point to a decision contract
(which can be the old one or one deployed by anybody else). The method is
decorated with modifier "onlyAuthorized()", which checks for the conditions
to either grant or reject the change of the referenced decision contract.
(This is the self-referential part of the governance mechanicsm.)

The "onlyAuthorized()" modified presently plays some tricks to allow for
simulating the passage of time in the development environment, and after
two rounds of interactions (controlled by the "passTime1" & "passTime2"
temporary variables), it uses the "require()" function that references a
method of the decision contract (line 63). The modifer makes use of the
decision contract by casting a local copy to the variable "how" from the
address of the decision-making contract that is presently in use (line 57).

In summary, this contract is allowed to change its governance mechanism,
and has placeholders for the laws constructs.

Details on "INhow_interface.sol":

This file is the interface of a smart contract that is expected to have a
certain behaviour; the contract is supposed to be interacted with via the
method "changeDecisionProcess()", which returns "true" or "false" based on
the contract's internal logic, which the developer needs to program. The
contract is meant to accept the address of another decision-making contract,
which is to represent the decision making contract that is to succeed the
present contract at a higher level caller contract. That is, there is a
"Laws" contract that is making reference to a given contract, and it needs
to ask this contreact for whether a _specific_ contract (known via its
address) has hte blessing to go ahead and be used. This is the mechanism of
self-referential governance changes!

A concrete implementation of this contract is to have two methods, the
"simplePing()" and "returnTrue()", which are to be used for debugging and
testing. (These will be removed when this contract is in production.)

Note that this contract does NOT require any code to be used in doing other
things from the context of the Laws contract acting as a caller (changing
and adding laws, etc), and this will be authored in the future.

Details on "INhow_default.sol":

This is the dictatorial decision contract (i.e. one payable address on the
blockchain can make decisions on changing the governance mechanism), and it
is the default for the founding of a "Laws" contract at deployment. It is
prototyped after the (abstract) "DecisionContract" and takes the incoming
address to establish as the dictator (stores it in its internal address
member "dictatorAddr").

It has the address member "allowedContractAddr", which is initially set to
the de facto invalid address of the Ethereum blockchain (the use of this
address is desribed below).

The most important method of this contract is "changeDecisionProcess()",
which is supposed to be called from a "Laws" contract with an argument the
address of the address of the candidate contract to succeed this one. The
mechanism for allowing this function to proceed goes through the modifier
"contractAuthorized()". This modified checks for whether the internal member
"allowedContractAddr" is the same as the proposed contract, and if so, it
will allow the method to proceed, and return "true" to the caller. The idea
is that the dictator will have to have set the "allowedContractAddr" ahead
of time, implying that they somehow approved for this _specific_ contract
to be the decision-making contract replacement. The dictator will have to
first know about the contract and its deployed address, and to approve the
change by interacting with this contract via the method "initiateChange()",
and giving it the already deployed successor contract's address. This method
is guarded by the modifier "onlyAuthorized()", which checks for whether the
address interacting with it is the address of the dictator that is stored
internally in "dictatorAddr".

This contract also implements the "simplePing()" method, which is meant to
emit a "Log()" (printable string on the console of the EVM) in development.
It implements the "returnTrue()" for debugging purposes (hack in line 69).

NOTE: This is a fragile dictatorship, which can imploed immediatelly after
the "Laws" contract has been instantiated, and therefore the dictator is
inclined to immediatelly change the "DecisionContract" instance to something
else. This is because there is a possibility that the blockchain will accept
the "address(0)" in an interaction via the method "changeDecisionProcess()",
as the "contractAuthorized" modifier may proceed with the "require()" logic,
given that the incoming "newDecisionContractAddr" is "address(0)". Then, the
decision-making process will always "revert" (in Ethereum state lingo), and
the "Laws" contract is now useless. To avoid this, the dictator can take the
address of the automatically deployed "DecisionContract_Default" at "Laws"
instantiation, and authorized it via "initiateChange()". The reason this
cannot be done at the original instantiation is that the constructor of the
decision making contract cannot have a self reference (like "this") before
there is an instance to refer to. We are stuck with an initially fragile
dictatorship!

These contracts will be modified before they are deployed. DO NOT USE!
IN 2023/06/27