Corrected serious bug with the "Laws" contract in the file "IN.sol". The decorator (modifier) "onlyAuthorized()" was passing to the "how" decision making contract the value of the internal variable "decideContractAddr". In essence, the address of the proposed contract was not used, and instead the decision making contract was seeing itself as the proposed contract, which would always return "true", and therefore would permit the change. This was a security flaw! The logic in the decision making contrac would always pass because of the caller ("Laws") not doing what it was supposed to.
To fix this, the change made to the "Laws" contract was with the modifier
to accept the incoming address of the proposed contract, such that it passes
it as argument to the decision making contract. This monumental change was
made alongside naming changes to make things more intuitive. The modifier
is now called "changeDecisionAuthorized( address )" and it now accepts an
address (of a contract) via the "_newGovernContractAddr". This name makes
more sense, as it now explicitly refers to governane as its role. The member
"decideContractAddr" is now "goverContractAddr".

The default (dictatorial) contract in "INhow_default.sol" was changed to
use a self reference at the constructor; apparently, and according to my
dear ChatGPT, this is possible in Solidity. By doing this, there is no need
to reset the value stored in member "allowedContractAddr" (which is used
for transitioning this governance contract), and the dictatorship is no
longer fragile. This also saves on gas by skipping this assignment, for what
it is worth. Extensive diagnostics were added to this contract, with the
new emitter "LogAddress()".
IN 2023/07/19