Added a method "notifyOfProposedLaw()" to the governance contract interface to allow for controlling how a law is proposed, but also to allow for some decision-making internals (such as voting) to be established for a given law. The method returns "true" to indicate success (whatever that may mean in the future). Some old notes were cleaned.
The dictatorial governance contract "INhow_default.sol" was augmented to
have the "notifyOfProposedLaw()" method, where it is notified that a law is
proposed, and it returns "true". (Why not? A dictator may well allow anyone
to propose.)

Changed the map in the "Laws" contract to use a contract address for the
key; in this way a law is addressed in the map by a unique, unambiguous
identity. The member of the struct "associatedContract" is not redundant,
but will be cleaned up later. Added the member "author" to store the address
of the proposer, and the 16-bit member "status", to use for sanity checks,
existence, and for other reasons (to disclose later). Added the method
"proposeLaw()" to the "Laws" contract, which takes an address of the pre-
deployed law contract. a string for the description. The method is guarded
by the modifier "checkProposedLaw()", which checks for (1) whether the law
has already been proposed or added and rejects if so, and (2) whether the
governance contract allows for this proposal to go through. The second check
happens by invoking the method "notifyOfProposedLaw()", with argument the
incoming address of the law contract. The first check happens by checking
the default of the EVM for a structure's member that is an integer, which
is "0" (should the law not exist in the map, the "require()" will fail).
This is why when the "proposeLaw()" method is allowed to succeed, the struct
will have the "status" member initialized to "1". In this way, a law that
has already been proposed will make the modifier fail.

Some augmentations were made to the "INhow_byvote.sol" (democracy) contract
to make it compatible with the changes made above, but this contract is not
supposed to work.
IN 2023/08/03