Added the method "addLaw()" to the "Laws" contract, which is supposed to check the governance contract for approving a _specific_ law. It is guarded by the modifier "checkAddingLaw()". On success, it changes the "status" member of the law struct to "2", to indicate that this law is approved in the laws of this country. The modifier "checkAddingLaw()" first checks that the associated law contract address is already a proposed law (checks the "status" member for being "1"), and then ask the governance contract for whether this _specific_ law has been approved. (NOTE: unapproved laws now remain in the map, and a clean-up method may be added in the future.)
The interface to the governance contract "INhow_Interface.sol" has been
augmented with the function "notifyOfApprovingLaw()", which is a method that
is meant to ask the governance contract for whether a _specific_ law has
been approved. (The governance contract could have kept track of what laws
have been proposed, but it does not have to.)

The "dictatorship" contract in "INhow_default.sol" has been augmented with
the method "notifyOfApprovingLaw()" as required by the parent interface. It
disregards the incoming law contract address, but checks for whether this
interaction is performed by the internally stored "dictarorAddr", and if so,
it returns "true". (Essentially, an interaction by the dictator approves
any law.)

The (non-functional) "democracy" contract has been augmented with the method
"notifyOfApprovingLaw()" such that it compiles; it always rejects laws by
returing false (for now).
IN 2023/08/07