Augmented the interface of the governane contract "INhow_Interface.sol" to accept the address "motionerAddr" in all of its methods. This allows the governance contract to know who (a person or other contract) is performing the interaction; this can guard from a lot of things.
The "Laws" contract in "IN.sol" was modified to use the "msg.sender" address
to the methods of its associated governance contract.

The dictatorship contract "INhow_default.sol" was modified to adhere to the
changes made in the interface from which it is prototyped. In the case of
being notified to allow for an addition of a law contract, it uses this
incoming "motioner" address to test against its internal "dictatorAddr" such
that only the dictator can approve when a "Laws" contract performs this
interaction. (There is not security issue here with this function being
called arbitrarily with the dictator address as the motioner, as it only
matters when a Laws contract receives its bool outcome, which only matters
when invoked from a "Laws" contract.)

The placeholder for a "democracy" contract "INhow_byvote.sol" was modified
such that it compiles; it is not intended to actually work.
IN 2023/08/09