Augmented the interface of the governance contract to have two methods for first proposing and then approving the removal of a specific law. Both of the methods need the law's contract address and the motioner address.
The dictatorship contract in "INhow_default.sol" was augmented to have the
law removal functions dictated by the interface from which it is prototyped.
The methods work like the pair of methods used for approving laws, with
essentially identical functionality.

The placeholder of a "democracy" contract in "INhow_byvote.sol" was modified
to compiler with the latest changes in the interface.

The "Laws" contract was augmented to remove laws in two steps, first by
proposing the law and subsequently by changing the "status" member of its
corresponding struct in the map to "3". This arbitrary number is to be
interpreted as the law being inactive, while the struct remains in the map
as historical record.
IN 2023/08/09