Added two Solidity smart contracts to examplify how a "token" is minted and managed on the Ethereum blockchain, potentially for use in a voting smart contract that uses either a "sigle vote" or a "weighted vote". Two files were added:
- EMPtoken.sol (an ERC-20 compliant token smart contract book)
- EMPunittoken.sol (a "membership token" smart contract book)

The ERC-20 compliant smart contract has a built-in "faucet" style minting
and burning mechanism; this is necessary here so that this can be used as
an example. (Generally, the creation of tokens is separated from the token
storage, in the sense that the ERC-20 smart contract will need to expose
appropriate methods that increase (and potentially decrease) the balance
of the accounts. Such a method should be programmed to guard from an attack
by enforcing appropraite logic -which may use other smart contracts- to do
token "minting" and "burning".) Warning: do not use this smart contract for
anything serious. On-chain "Transfer" events are emitted when minting,
burning, or transfering of tokens takes place to allow for other mechanisms
to be triggered on and off chain.

The "membership token" smart contract uses minimal on-chain storage to allow
for a certain Ethereum address ("account") to be granted a token, which is
referred to as "unit" here, because it does not have an assocaited quantity
of tokens. This contract is _not_ ERC-20 compliant. A minting and burning
faucet style mechanism was also added to make this example work. On-chain
events are emitted when minting, burning, and transfer of membership tokens
takes place.
IN 2023/11/26